Solved PT07Z
Find the longest path in a tree without being given the root.  Pick an arbitrary node to start with and run a dfs.  The farthest node from the start will be a leaf node guaranteed to be part of the actual max path, so run another dfs from that leaf to get the final answer.