Solved ACODE
Had to break out the pencil and paper for this one.  Tons of corner cases to think about.

This solution examines each character in the input string, conceptually (but not literally) generating a tree as it goes.  The digit at index i will generate one or more nodes in the tree with depth of i.  The number of nodes generated depends on the number of nodes of different types (ones, twos, others) at the depth above it.  The final value is the number of leaf nodes at the maximum depth.

This would probably be a lot easier to read and write recursively, but I was intent on solving it with dp for the sake of practice.