Solved ANARC05B
The problem is tagged dynamic programming and binary search, but actually requires neither.

This solution just finds the intersections ("marks") by sticking the values of each array into a map and finding the collisions, then calculates the intervals up to those intersection points for each array, then greedily walks the intervals to get the maximum sum.